Added FMEA object which stores a list of components
Changed LoadPath to now read the data from the XML files to create a new object for each found in the XML file.
This is currently stored in a list of HiP_HOPSResults, to be combined later.

Added the following methods to the program that read in the data from the XML files and create the respective object:
ReadHiPHOPSResults
ReadFaultTree
ReadCutSets
ReadOutputDeviation
ReadChildren - (List of Nodes)
ReadOr
ReadAnd
ReadFMEA
ReadEvent
ReadEffect