Added a static List of Dictionaries which stores each unique basic event and its previous ID for each different HiP-HOPSResults.
Fixed ReadComponent to now correctly add the events to the component.
Fixed ReadEvent to now correctly add the effects to the event.

Changed ReadEvent to check if the current HiP-HOPSResults contains previous ID of the event to be read in.

Added methods to begin combining duplicate Basic events and components.